requirements.txt
cs50
Flask
Flask-Session
requests
sqlite3
Python 3
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib